+++ library/core/src/main/java/com/google/android/exoplayer2/upstream/ContentDataSource.java
@@ -21,10 +21,18 @@ import static java.lang.Math.min;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.content.res.AssetFileDescriptor;
+import android.media.ApplicationMediaCapabilities;
+import android.media.MediaFeature;
+import android.media.MediaFormat;
 import android.net.Uri;
+import android.os.Bundle;
+import android.provider.MediaStore;
+import androidx.annotation.DoNotInline;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.PlaybackException;
+import com.google.android.exoplayer2.util.Util;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -70,10 +78,17 @@ public final class ContentDataSource extends BaseDataSource {
       this.uri = uri;
 
       transferInitializing(dataSpec);
-      AssetFileDescriptor assetFileDescriptor = resolver.openAssetFileDescriptor(uri, "r");
+
+      Bundle providerOptions = new Bundle();
+      if (Util.SDK_INT >= 31) {
+        Api31.disableTranscoding(providerOptions);
+      }
+
+      AssetFileDescriptor assetFileDescriptor =
+          resolver.openTypedAssetFileDescriptor(uri, /* mimeType= */ "*/*", providerOptions);
       this.assetFileDescriptor = assetFileDescriptor;
       if (assetFileDescriptor == null) {
-        // openAssetFileDescriptor returns null if the provider recently crashed.
+        // openTypedAssetFileDescriptor returns null if the provider recently crashed.
         throw new ContentDataSourceException(
             new IOException("Could not open file descriptor for: " + uri),
             PlaybackException.ERROR_CODE_IO_UNSPECIFIED);
@@ -203,6 +218,23 @@ public final class ContentDataSource extends BaseDataSource {
           transferEnded();
         }
       }
+    }
+  }
+
+  @RequiresApi(31)
+  private static final class Api31 {
+
+    @DoNotInline
+    public static void disableTranscoding(Bundle providerOptions) {
+      ApplicationMediaCapabilities mediaCapabilities =
+          new ApplicationMediaCapabilities.Builder()
+              .addSupportedVideoMimeType(MediaFormat.MIMETYPE_VIDEO_HEVC)
+              .addSupportedHdrType(MediaFeature.HdrType.DOLBY_VISION)
+              .addSupportedHdrType(MediaFeature.HdrType.HDR10)
+              .addSupportedHdrType(MediaFeature.HdrType.HDR10_PLUS)
+              .addSupportedHdrType(MediaFeature.HdrType.HLG)
+              .build();
+      providerOptions.putParcelable(MediaStore.EXTRA_MEDIA_CAPABILITIES, mediaCapabilities);
     }
   }
 }